Moved playground pages to their own directory. Add Wallets page. Rename some of the pages for more consistency.
