--- playground/Components/Components.html
@@ -1,154 +1,0 @@
-<!doctype html>
-<html class="no-js" lang="">
-
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <title>Components</title>
-    <meta name="description" content="">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-</head>
-
-<body>
-    <header></header>
-
-    <form class="merchant-checkout__form" method="post">
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>Would you like to donate?</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="donation-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>MB Way</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="mbway-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>ACH</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="ach-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>AfterPay</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="afterpay-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>FacilyPay 3x</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="facilypay_3x-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>FacilyPay 4x</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="facilypay_4x-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>Sepa</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="sepa-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>Qiwi Wallet</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="qiwi-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>Vipps</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="vipps-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>Giropay</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="giropay-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>blik</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="blik-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <script async src="https://pay.google.com/gp/p/js/pay.js"></script>
-            <div class="merchant-checkout__payment-method__header">
-                <h2>Google Pay</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="googlepay-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>Paypal</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="paypal-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method merchant-checkout__payment-method--hidden" id="applepay-field">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>Apple Pay</h2>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>Redirect</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="redirect-field"></div>
-            </div>
-        </div>
-    </form>
-
-    <script type="text/javascript">
-        window.htmlPages = <%= JSON.stringify(htmlWebpackPlugin.htmlPages) || ''%>;
-    </script>
-</body>
-</html>
