--- playground/IssuerLists/IssuerLists.html
@@ -1,102 +1,0 @@
-<!doctype html>
-<html class="no-js" lang="">
-
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <title>Checkout Components</title>
-    <meta name="description" content="">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-</head>
-
-<body>
-    <header></header>
-
-    <form class="merchant-checkout__form" method="post">
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>iDEAL - NL 🇳🇱</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="ideal-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>BillDesk Online - IN 🇮🇳</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="billdesk_online-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>BillDesk Wallet - IN 🇮🇳</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="billdesk_wallet-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>PayU CashCard - IN 🇮🇳</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="payu_cc-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>PayU NetBanking - IN 🇮🇳</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="payu_nb-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>Dotpay - PL 🇵🇱</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="dotpay-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>Entercash - FI 🇫🇮</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="entercash-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method">
-            <div class="merchant-checkout__payment-method__header">
-                <h2>OpenBanking - GB 🇬🇧</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="openbanking_UK-field"></div>
-            </div>
-        </div>
-
-        <div class="merchant-checkout__payment-method" >
-            <div class="merchant-checkout__payment-method__header">
-                <h2>MolPay - MY 🇲🇾</h2>
-            </div>
-            <div class="merchant-checkout__payment-method__details">
-                <div class="molpay-field"></div>
-            </div>
-        </div>
-    </form>
-
-    <script type="text/javascript">
-        window.htmlPages = <%= JSON.stringify(htmlWebpackPlugin.htmlPages) || '' %>;
-    </script>
-</body>
-</html>
