+++ playground/pages/Cards/Cards.html
@@ -1,0 +1,197 @@
+<!doctype html>
+<html class="no-js" lang="">
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="x-ua-compatible" content="ie=edge">
+    <title>Adyen Web | Cards</title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+</head>
+
+<body>
+    <header></header>
+
+    <form class="merchant-checkout__form" method="post">
+        <div class="merchant-checkout__payment-method ">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>Stored card</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div class="oneclick-field"></div>
+            </div>
+        </div>
+
+        <div class="merchant-checkout__payment-method">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>Card</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div class="card-field"></div>
+            </div>
+        </div>
+
+        <div class="merchant-checkout__payment-method">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>Bancontact</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div class="bancontact-field"></div>
+            </div>
+        </div>
+
+        <div class="merchant-checkout__payment-method">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>Card with Billing Address</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div class="card-avs-field"></div>
+            </div>
+        </div>
+
+        <div class="merchant-checkout__payment-method">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>Card with Korean Authentication</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div class="card-kcp-field"></div>
+            </div>
+        </div>
+    </form>
+
+    <div class="info">
+        <h2>3DS2:</h2>
+        <table>
+            <thead>
+                <tr>
+                    <th>Card Type</th>
+                    <th>Card Number</th>
+                </tr>
+            </thead>
+            <tbody>
+                <tr>
+                    <td>amex</td>
+                    <td>3714 4963 5398 431</td>
+                </tr>
+                <tr>
+                    <td>Cartes Bancaires</td>
+                    <td>4035 5014 2814 6300</td>
+                </tr>
+                <tr>
+                    <td>Diners</td>
+                    <td>3056 9309 0259 04</td>
+                </tr>
+                <tr>
+                    <td>Discover</td>
+                    <td>6011 1111 1111 1117</td>
+                </tr>
+                <tr>
+                    <td>JCB</td>
+                    <td>3566 1111 1111 1113</td>
+                </tr>
+                <tr>
+                    <td>mc</td>
+                    <td>5454 5454 5454 5454</td>
+                </tr>
+                <tr>
+                    <td>UnionPay</td>
+                    <td>6212 3456 7890 1232</td>
+                </tr>
+                <tr>
+                    <td>visa</td>
+                    <td>4917 6100 0000 0000</td>
+                </tr>
+            </tbody>
+        </table>
+
+        <h2>AVS:</h2>
+        <table>
+            <thead>
+                <tr>
+                    <th>Card Number</th>
+                    <th>Card Type</th>
+                    <th>House Number</th>
+                    <th>Address</th>
+                    <th>ZIP Code</th>
+                    <th>Country</th>
+                </tr>
+            </thead>
+            <tbody>
+                <tr>
+                    <td>374251021090003</td>
+                    <td>amex</td>
+                    <td>1</td>
+                    <td>Infinite Loop Cupertino, CA.</td>
+                    <td>95014</td>
+                    <td>US</td>
+                </tr>
+                <tr>
+                    <td>374251018720018</td>
+                    <td>amex</td>
+                    <td>1600</td>
+                    <td>Pennsylvania Ave NW Washington, DC.</td>
+                    <td>20500</td>
+                    <td>US</td>
+                </tr>
+                <tr>
+                    <td>374101012180018</td>
+                    <td>amex</td>
+                    <td>10</td>
+                    <td>Downing Street, London.</td>
+                    <td>SW1A 2AA</td>
+                    <td>GB</td>
+                </tr>
+                <tr>
+                    <td>374251033270007</td>
+                    <td>amex</td>
+                    <td>8-10</td>
+                    <td>Broadway, Westminster, London.</td>
+                    <td>SW1H 0BG</td>
+                    <td>GB</td>
+                </tr>
+                <tr>
+                    <td>5500000000000004</td>
+                    <td>mc</td>
+                    <td>1600</td>
+                    <td>Pennsylvania Ave NW Washington, DC.</td>
+                    <td>20500</td>
+                    <td>US</td>
+                </tr>
+                <tr>
+                    <td>5555555555554444</td>
+                    <td>mc</td>
+                    <td>10</td>
+                    <td>Downing Street, London.</td>
+                    <td>SW1A 2AA</td>
+                    <td>GB</td>
+                </tr>
+                <tr>
+                    <td>4400000000000008</td>
+                    <td>visa</td>
+                    <td>1</td>
+                    <td>Infinite Loop Cupertino, CA.</td>
+                    <td>95014</td>
+                    <td>US</td>
+                </tr>
+                <tr>
+                    <td>4444333322221111</td>
+                    <td>visa</td>
+                    <td>8-10</td>
+                    <td>Broadway, Westminster, London.</td>
+                    <td>SW1H 0BG</td>
+                    <td>GB</td>
+                </tr>
+            </tbody>
+        </table>
+
+        <p>
+            <a href="https://docs.adyen.com/development-resources/test-cards/test-card-numbers/#test-3d-secure-2-authentication" target="_blank">
+                More information →
+            </a>
+        </p>
+    </div>
+
+    <script type="text/javascript">
+    window.htmlPages = <%= JSON.stringify(htmlWebpackPlugin.htmlPages) || '' %>;
+</script>
+</body>
+</html>
