+++ playground/pages/Dropin/Dropin.html
@@ -1,0 +1,23 @@
+<!doctype html>
+<html class="no-js" lang="en">
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="x-ua-compatible" content="ie=edge">
+    <title>Adyen Web | Drop-in</title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+</head>
+
+<body>
+    <header></header>
+
+    <div class="merchant-checkout__form">
+        <div id="dropin-container"></div>
+    </div>
+
+    <script src="https://pay.google.com/gp/p/js/pay.js"></script>
+    <script type="text/javascript">
+        window.htmlPages = <%= JSON.stringify(htmlWebpackPlugin.htmlPages) || '' %>;
+    </script>
+</body>
+</html>
