+++ playground/pages/GiftCards/GiftCards.html
@@ -1,0 +1,112 @@
+<!doctype html>
+<html class="no-js" lang="en">
+
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="x-ua-compatible" content="ie=edge">
+    <title>Adyen Web | Gift Cards</title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+</head>
+
+<body class="giftcard-demo">
+    <header></header>
+
+    <div class="merchant-checkout__form">
+
+        <div class="merchant-checkout__payment-method">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>Giftcard</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div id="genericgiftcard-container"></div>
+            </div>
+        </div>
+
+        <div class="merchant-checkout__payment-method">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>Applied Giftcard</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div id="genericgiftcard-result-container"></div>
+            </div>
+        </div>
+    </div>
+
+
+    <div class="info">
+
+        <table>
+            <thead>
+                <tr>
+                    <th>Card Type</th>
+                    <th>Card Number</th>
+                    <th>PIN</th>
+                </tr>
+            </thead>
+            <tbody>
+                <tr>
+                    <td>Gall & Gall Card</td>
+                    <td>60643650100000000000</td>
+                    <td>73737</td>
+                </tr>
+                <tr>
+                    <td>Baby Gift Card</td>
+                    <td>60643622000000000000</td>
+                    <td>73737</td>
+                </tr>
+                <tr>
+                    <td>Gift Card</td>
+                    <td>62805011000000000000</td>
+                    <td>73737</td>
+                </tr>
+                <tr>
+                    <td>Kado Wereld</td>
+                    <td>60643625100000000000</td>
+                    <td>73737</td>
+                </tr>
+                <tr>
+                    <td>Entertainment Card</td>
+                    <td>60643611000000000000</td>
+                    <td>73737</td>
+                </tr>
+                <tr>
+                    <td>Plastix</td>
+                    <td>4010100000000000000</td>
+                    <td>73737</td>
+                </tr>
+                <tr>
+                    <td>Webshop Giftcard</td>
+                    <td>60643620700000000000</td>
+                    <td>73737</td>
+                </tr>
+                <tr>
+                    <td>Leasure Giftcard</td>
+                    <td>60643622800000000000</td>
+                    <td>73737</td>
+                </tr>
+                <tr>
+                    <td>VVV Giftcard</td>
+                    <td>6064364240000000000</td>
+                    <td>737373</td>
+                </tr>
+                <tr>
+                    <td>GiftForYou (Bloemen)</td>
+                    <td>60643647103300000000</td>
+                    <td>737373</td>
+                </tr>
+            </tbody>
+        </table>
+
+        <p>
+            <a href="https://docs.adyen.com/development-resources/test-cards/test-card-numbers#gift-cards" target="_blank">
+                More information →
+            </a>
+        </p>
+    </div>
+
+    <script type="text/javascript">
+    window.htmlPages = <%= JSON.stringify(htmlWebpackPlugin.htmlPages) || '' %>;
+</script>
+</body>
+</html>
