+++ playground/pages/OpenInvoices/OpenInvoices.html
@@ -1,0 +1,48 @@
+<!doctype html>
+<html class="no-js" lang="">
+
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="x-ua-compatible" content="ie=edge">
+    <title>Adyen Web | Open Invoices</title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+</head>
+
+<body>
+    <header></header>
+
+    <form class="merchant-checkout__form" method="post">
+        <div class="merchant-checkout__payment-method">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>AfterPay</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div class="afterpay-field"></div>
+            </div>
+        </div>
+
+        <div class="merchant-checkout__payment-method">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>FacilyPay 3x</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div class="facilypay_3x-field"></div>
+            </div>
+        </div>
+
+        <div class="merchant-checkout__payment-method">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>RatePay</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div class="ratepay-field"></div>
+            </div>
+        </div>
+    </form>
+
+    <script type="text/javascript">
+        window.htmlPages = <%= JSON.stringify(htmlWebpackPlugin.htmlPages) || '' %>;
+    </script>
+</body>
+</html>
