+++ playground/pages/QRCodes/QRCodes.html
@@ -1,0 +1,51 @@
+<!doctype html>
+<html class="no-js" lang="en">
+
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="x-ua-compatible" content="ie=edge">
+    <title>Adyen Web | QR Codes</title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+</head>
+
+<body class="qrcode-demo">
+    <header></header>
+
+    <form class="merchant-checkout__form" method="post">
+
+        <div class="merchant-checkout__payment-method">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>WeChat Pay QR</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div id="wechatpayqr-container"></div>
+            </div>
+        </div>
+
+
+        <div class="merchant-checkout__payment-method">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>Bancontact App QR</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div id="bcmcqr-container"></div>
+            </div>
+        </div>
+
+        <div class="merchant-checkout__payment-method">
+            <div class="merchant-checkout__payment-method__header">
+                <h2>Swish</h2>
+            </div>
+            <div class="merchant-checkout__payment-method__details">
+                <div id="swish-container"></div>
+            </div>
+        </div>
+
+    </form>
+
+    <script type="text/javascript">
+        window.htmlPages = <%= JSON.stringify(htmlWebpackPlugin.htmlPages) || '' %>;
+    </script>
+</body>
+</html>
