+++ playground/pages/QRCodes/QRCodes.scss
