+++ playground/pages/SecuredFields/securedFields-fancy.config.js
