+++ playground/pages/SecuredFields/securedFields.config.js
