+++ playground/pages/Vouchers/Vouchers.html
@@ -1,0 +1,59 @@
+<!doctype html>
+<html class="no-js" lang="en">
+
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="x-ua-compatible" content="ie=edge">
+    <title>Adyen Web | Vouchers</title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+</head>
+
+<body class="voucher-demo">
+    <header></header>
+
+    <div class="merchant-checkout__form">
+        <h2>Multibanco</h2>
+        <div id="multibanco-result-container"></div>
+    </div>
+
+    <div class="merchant-checkout__form">
+        <h2>Boleto Bancario</h2>
+        <div id="boleto-input-container" class="demo-voucher-input"></div>
+        <div id="boleto-result-container"></div>
+    </div>
+
+    <div class="merchant-checkout__form">
+        <h2>OXXO</h2>
+        <div id="oxxo-result-container"></div>
+    </div>
+
+    <div class="merchant-checkout__form">
+        <h2>Dragonpay</h2>
+        <div id="dragonpay-input-container" class="demo-voucher-input"></div>
+        <div id="dragonpay-result-container"></div>
+    </div>
+
+    <div class="merchant-checkout__form">
+        <h2>Doku</h2>
+        <div id="doku-input-container" class="demo-voucher-input"></div>
+        <div id="doku-result-container"></div>
+    </div>
+
+    <div class="merchant-checkout__form">
+        <h2>Econtext</h2>
+        <div id="econtext-input-container" class="demo-voucher-input"></div>
+        <div id="econtext-result-container"></div>
+    </div>
+
+    <div class="merchant-checkout__form">
+        <h2>Econtext 7-Eleven</h2>
+        <div id="econtext-seven-eleven-input-container" class="demo-voucher-input"></div>
+        <div id="econtext-seven-eleven-result-container"></div>
+    </div>
+
+    <script type="text/javascript">
+        window.htmlPages = <%= JSON.stringify(htmlWebpackPlugin.htmlPages) || '' %>;
+    </script>
+</body>
+</html>
