+++ playground/pages/Vouchers/Vouchers.scss
@@ -1,0 +1,12 @@
+@import 'src/style';
+
+.voucher-demo h2 {
+    margin: 16px 0;
+}
+
+.demo-voucher-input {
+    background: $color-white;
+    padding: 24px;
+    border: 1px solid $color-gray;
+    margin-bottom: 16px;
+}
